+++ backend/src/main/java/com/earntogether/qlysotietkiem/dto/TermInsertDTO.java
@@ -3,11 +3,11 @@ package com.earntogether.qlysotietkiem.dto;
 import jakarta.validation.constraints.*;
 import lombok.Builder;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Builder
 public record TermInsertDTO(
         @PositiveOrZero(message = "Invalid number of term months") int monthsOfTerm,
         @PositiveOrZero(message = "Invalid interest rate") double interestRate,
-        @NotNull(message = "Invalid minimum deposit amount") BigInteger minDeposit
+        @NotNull(message = "Invalid minimum deposit amount") BigDecimal minDeposit
 ) {}
