+++ backend/src/main/java/com/earntogether/qlysotietkiem/model/WithdrawalSlipModel.java
@@ -2,12 +2,12 @@ package com.earntogether.qlysotietkiem.model;
 
 import lombok.Builder;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Builder
 public record WithdrawalSlipModel(String id, int customerCode,
                                   int passbookCode,
                                   int type, LocalDate withdrawalDate,
-                                  BigInteger money
+                                  BigDecimal money
 ) {}
