+++ backend/src/main/java/com/earntogether/qlysotietkiem/utils/converter/TermConverter.java
@@ -3,7 +3,7 @@ package com.earntogether.qlysotietkiem.utils.converter;
 import com.earntogether.qlysotietkiem.dto.TermInsertDTO;
 import com.earntogether.qlysotietkiem.entity.Term;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class TermConverter {
     public static Term convertDTOtoEntity(TermInsertDTO termInsertDto, int type) {
@@ -14,7 +14,7 @@ public class TermConverter {
         int daysWithdrawn = 15;
         // Chỉ áp dụng cho không kỳ hạn
         var minAdditionalDeposit = monthsOfTerm == 0 ?
-                BigInteger.valueOf(100000) : BigInteger.valueOf(0);
+                BigDecimal.valueOf(100000) : BigDecimal.valueOf(0);
         // Mặc định số tháng của Không kì hạn(monthsOfTerm) là 1
         return Term.builder()
                 .type(type)
